pyproject.toml
[project]
name = "argus"
version = "0.1.0"
description = "Automated Real-time Global Universal Surveillance"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "alembic>=1.14.1",
    "bcrypt>=4.2.1",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "jwt>=1.3.1",
    "passlib>=1.7.4",
    "pathlib>=1.0.1",
    "psutil>=7.0.0",
    "pydantic-settings>=2.8.0",
    "pyjwt>=2.10.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "starlette>=0.45.3",
    "typing-extensions>=4.12.2",
    "uvicorn[standard]>=0.34.0",
]
